# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEN Signal Intelligence API",
    "description": "## Signal Intelligence Engine\n\nOMEN transforms market events into structured probability signals.\n\n### Authentication\n\nAll API endpoints require authentication via API key.\nInclude the `X-API-Key` header in all requests.\n\n**Public endpoints (no auth required):**\n- `/health/*` - Health checks\n- `/docs`, `/redoc` - API documentation\n- `/metrics` - Prometheus metrics\n\n### What OMEN provides\n- Probability assessment with confidence bounds\n- Signal validation and quality scoring\n- Geographic and temporal context\n- Full evidence chain and traceability\n\n### What OMEN does NOT provide\n- Impact assessment (severity, delay, cost)\n- Decision steering (temporal proximity, relevance rank, high-confidence flags)\n- Recommendations or advice\n- Risk quantification\n\nImpact assessment is the responsibility of downstream systems.",
    "version": "2.0.0",
    "x-logo": {
      "url": "https://omen.io/logo.png",
      "altText": "OMEN Logo"
    },
    "contact": {
      "name": "OMEN API Support",
      "url": "https://github.com/omen/omen",
      "email": "support@omen.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/api/v1/signals/": {
      "get": {
        "tags": [
          "Signals"
        ],
        "summary": "List recent signals",
        "description": "List recent signals with pagination.\n\n**Query Parameters:**\n- `limit`: Max signals to return (default: 100, max: 1000)\n- `offset`: Pagination offset (default: 0)\n- `since`: Only return signals after this timestamp (ISO 8601)\n\n**Example Request:**\n```\nGET /api/v1/signals/?limit=10&since=2026-01-01T00:00:00Z\n```\n\n**Response includes:**\n- List of redacted signals\n- Pagination metadata (total, limit, offset)",
        "operationId": "list_signals_api_v1_signals__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "description": "Maximum number of signals to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of signals to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only return signals after this timestamp",
              "title": "Since"
            },
            "description": "Only return signals after this timestamp"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of signals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Signals Api V1 Signals  Get"
                },
                "example": {
                  "items": [
                    {
                      "signal_id": "sig-poly-12345",
                      "signal_type": "MARKET_MOVEMENT",
                      "title": "Polymarket: US Election probability shift",
                      "description": "Significant movement in election prediction markets",
                      "probability": 0.65,
                      "confidence": 0.88,
                      "keywords": [
                        "election",
                        "politics",
                        "prediction"
                      ],
                      "source": "polymarket",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "has_more": false,
                  "cursor": null
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/multi-source/sources": {
      "get": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "List Sources",
        "description": "List all registered signal sources and their status.\n\nReturns health status of each source:\n- Polymarket: Prediction market data\n- AIS: Ship tracking, port congestion\n- Weather: Storm alerts, sea conditions\n- Freight: Container rates, capacity",
        "operationId": "list_sources_api_v1_multi_source_sources_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceHealthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/multi-source/signals": {
      "get": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "Get Multi Source Signals",
        "description": "Get signals from multiple sources.\n\nSources:\n- polymarket: Prediction market signals\n- ais: Port congestion, chokepoint delays\n- weather: Storm alerts, sea conditions\n- freight: Rate spikes, capacity alerts\n\nSignals are combined and sorted by timestamp.",
        "operationId": "get_multi_source_signals_api_v1_multi_source_signals_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit_per_source",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 20,
              "title": "Limit Per Source"
            }
          },
          {
            "name": "sources",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated source names (e.g., 'ais,weather,freight')",
              "title": "Sources"
            },
            "description": "Comma-separated source names (e.g., 'ais,weather,freight')"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiSourceResponse"
                },
                "example": {
                  "items": [
                    {
                      "signal_id": "sig-poly-12345",
                      "signal_type": "MARKET_MOVEMENT",
                      "title": "Polymarket: US Election probability shift",
                      "description": "Significant movement in election prediction markets",
                      "probability": 0.65,
                      "confidence": 0.88,
                      "keywords": [
                        "election",
                        "politics",
                        "prediction"
                      ],
                      "source": "polymarket",
                      "timestamp": "2026-02-01T10:00:00Z"
                    }
                  ],
                  "has_more": false,
                  "cursor": null
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/multi-source/sources/{source_name}": {
      "patch": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "Update Source",
        "description": "Update a signal source configuration.\n\nSet enabled=true to enable, enabled=false to disable.",
        "operationId": "update_source_api_v1_multi_source_sources__source_name__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "source_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Name"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Source Api V1 Multi Source Sources  Source Name  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/multi-source/sources/{source_name}/signals": {
      "get": {
        "tags": [
          "Multi-Source Intelligence"
        ],
        "summary": "Get Single Source Signals",
        "description": "Get signals from a specific source.",
        "operationId": "get_single_source_signals_api_v1_multi_source_sources__source_name__signals_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "source_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Name"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Api Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MultiSourceResponse": {
        "properties": {
          "signals": {
            "items": {
              "$ref": "#/components/schemas/MultiSourceSignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "by_source": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "By Source"
          },
          "fetch_time_ms": {
            "type": "number",
            "title": "Fetch Time Ms"
          }
        },
        "type": "object",
        "required": [
          "signals",
          "total",
          "by_source",
          "fetch_time_ms"
        ],
        "title": "MultiSourceResponse",
        "description": "Response from multi-source fetch."
      },
      "MultiSourceSignal": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "probability": {
            "type": "number",
            "title": "Probability"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locations"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "source_metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Metrics"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "source",
          "title",
          "description",
          "probability",
          "keywords",
          "locations",
          "timestamp"
        ],
        "title": "MultiSourceSignal",
        "description": "Signal from multi-source aggregation."
      },
...
      "SourceHealthResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceInfo"
            },
            "type": "array",
            "title": "Sources"
          },
          "healthy_count": {
            "type": "integer",
            "title": "Healthy Count"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "sources",
          "healthy_count",
          "total_count"
        ],
        "title": "SourceHealthResponse",
        "description": "Health status of all sources."
      },
...
      "SourceInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "unknown"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enabled",
          "priority",
          "weight"
        ],
        "title": "SourceInfo",
        "description": "Information about a signal source."
      },
      "SourceUpdateRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "Enable or disable the source"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "SourceUpdateRequest",
        "description": "Request to update a source's configuration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token (optional, if enabled)"
      }
    }
  },
  "servers": [
    {
      "url": "https://api.omen.io",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.omen.io",
      "description": "Staging server"
...
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}